# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "TestApp API",
        "contact": {
            "name": "Xamelllion",
            "url": "t.me/xamelllion",
            "email": "me@xamelllion.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://mit-license.org/"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/ping": {
            "get": {
                "description": "get status of services",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service"
                ],
                "summary": "Get status of services",
                "operationId": "get-status-of-services",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ServicesInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/domain.ServicesInfo"
                        }
                    }
                }
            }
        },
        "/api/repositories/info": {
            "get": {
                "description": "get repository information by GitHub repository URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repo"
                ],
                "summary": "Get repository by GitHub URL",
                "operationId": "get-repo-by-owner-repo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GitHub repository URL",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Repository"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/subscriptions": {
            "get": {
                "description": "Get all repository subscriptions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get subscriptions",
                "operationId": "get-subscriptions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Subscription"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a subscription to an existing GitHub repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Subscribe to a repository",
                "operationId": "subscribe-to-repository",
                "parameters": [
                    {
                        "description": "Repository subscription",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.SubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/subscriptions/info": {
            "get": {
                "description": "Get GitHub information for every subscribed repository",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get subscribed repository information",
                "operationId": "get-subscribed-repository-info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Repository"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/subscriptions/{owner}/{repo}": {
            "delete": {
                "description": "Delete an existing repository subscription",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Unsubscribe from a repository",
                "operationId": "unsubscribe-from-repository",
                "parameters": [
                    {
                        "type": "string",
                        "example": "octocat",
                        "description": "GitHub repository owner",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Hello-World",
                        "description": "GitHub repository name",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRepository.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.GetRepository.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some error"
                }
            }
        },
        "controller.SubscribeRequest": {
            "type": "object",
            "properties": {
                "owner": {
                    "type": "string",
                    "example": "octocat"
                },
                "repo": {
                    "type": "string",
                    "example": "Hello-World"
                }
            }
        },
        "domain.PingStatus": {
            "type": "string",
            "enum": [
                "up",
                "down"
            ],
            "x-enum-varnames": [
                "PingStatusUp",
                "PingStatusDown"
            ]
        },
        "domain.Repository": {
            "type": "object",
            "properties": {
                "create_date": {
                    "type": "string",
                    "example": "2026-03-03T16:45:04Z"
                },
                "description": {
                    "type": "string",
                    "example": "xamelllion's repo"
                },
                "forks": {
                    "type": "integer",
                    "example": 2
                },
                "name": {
                    "type": "string",
                    "example": "xamelllion"
                },
                "stars": {
                    "type": "integer",
                    "example": 122
                }
            }
        },
        "domain.ServiceStatus": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.PingStatus"
                }
            }
        },
        "domain.ServicesInfo": {
            "type": "object",
            "properties": {
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ServiceStatus"
                    }
                },
                "status": {
                    "$ref": "#/definitions/domain.ServicesStatus"
                }
            }
        },
        "domain.ServicesStatus": {
            "type": "string",
            "enum": [
                "ok",
                "degraded"
            ],
            "x-enum-varnames": [
                "ServicesStatusOk",
                "ServicesStatusDegraded"
            ]
        },
        "domain.Subscription": {
            "type": "object",
            "properties": {
                "owner": {
                    "type": "string",
                    "example": "octocat"
                },
                "repo": {
                    "type": "string",
                    "example": "Hello-World"
                }
            }
        }
    }
}